Summarizer (#3)
* Minor changes

* Minor changes

* feat: Update summary tree generation and prompts

The code changes include adding new functionality for generating a summary tree and prompts for summarizing each section node. This improves the summarization process by providing a hierarchical structure and clear instructions for summarizing each section.

* feat: Update filepath and add hierarchical summary dataset

The code changes in `catalog.yml` update the filepath for the `tree_cut` dataset and add a new dataset called `hierarchical_summary`. This improves the organization of the data and allows for better analysis and processing of the hierarchical summary information.

* feat: Update parameters and add safety settings for LLMEngine

The code changes in `conf/base/parameters_summarize.yml` update the `tree_cut` dataset filepath and add new parameters for the `llm_engine` configuration. This improves the organization of the data and allows for better analysis and processing of the hierarchical summary information. Additionally, the `llm_engine.py` file is updated to include safety settings for blocking harmful content. These settings specify the threshold for blocking dangerous content, hate speech, harassment, and sexually explicit content.

This commit message follows the established convention of using a prefix to indicate the type of change, followed by a concise and descriptive message.

* Update parameters

* Restore LLM engine gemini model

* Fix

* feat: Experiment flat query

* fix. Code cleaning

* [FEAT] Prompt Enginerring Optimisation

---------

Co-authored-by: Brendan L'OLLIVIER <blollivier@idmog.com>